requirements.txt
streamlit==1.31.1
langchain-groq==0.1.3
langchain==0.1.13
unstructured==0.12.2
sentence-transformers==2.6.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
pickle5==0.0.12
langchain-community==0.0.29